requirements.txt

flask
gunicorn
PyPDF2
google-auth
google-auth-oauthlib
google-api-python-client
